C++ code updated for OMNet++ 4.0, using omnetpp\_scripts\migrate\migratecpp.rb
git-svn-id: file:///home/rhornig/tmp/svn/inet-framework/branches/tmp-omnetpp40port@1508 539eec35-cc16-0410-b457-356aab9780e1

+++ NetworkInterfaces/MF80211/macLayer/CSMAMacLayer.cc
@@ -43,7 +43,7 @@ void CSMAMacLayer::initialize(int stage)
 
     if (stage == 0)
     {
-        queueLength = hasPar("queueLength") ? par("queueLength") : 0;
+        queueLength = hasPar("queueLength") ? (int)par("queueLength") : 0;
         EV << "queueLength = " << queueLength << endl;
 
         //subscribe for the information of the carrier sense
